requirements.txt
python-telegram-bot==20.7
requests>=2.32.3,<3.0
matplotlib>=3.9.0,<3.10
